### 2019-09-02 #44 @whedon
[PRE REVIEW]: ∂P: A Differentiable Programming System to Bridge Machine Learning and Scientific Computing
**Submitting author:** @MikeInnes (<a href="http://orcid.org/0000-0003-0788-0242">Michael Innes</a>)
**Repository:** <a href="https://github.com/MikeInnes/zygote-paper" target ="_blank">https://github.com/MikeInnes/zygote-paper</a>

Thanks for submitting your paper to JuliaCon @MikeInnes. The JuliaCon editor @vchuravy, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@MikeInnes if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
