+++ paper/paper.tex
@@ -4,7 +4,7 @@
 \setcounter{page}{1}
 
 \usepackage{hyperref}
-\usepackage[linesnumbered,ruled,vlined]{algorithm2e}
+% \usepackage[linesnumbered,ruled,vlined]{algorithm2e}
 \usepackage{graphicx}
 \usepackage[font=small,labelfont=bf]{caption}
 \usepackage{subcaption}
