### 2019-09-23 f6d4b40
Update paper.tex
