+++ paper/paper.tex
@@ -8,7 +8,7 @@
 \usepackage{graphicx}
 \usepackage[font=small,labelfont=bf]{caption}
 \usepackage{subcaption}
-\usepackage{todonotes}
+% \usepackage{todonotes}
 \usepackage{amsmath}
 \usepackage{etoolbox}
 \usepackage{comment}
