+++ paper/paper.tex
@@ -11,7 +11,7 @@
 % \usepackage{todonotes}
 \usepackage{amsmath}
 \usepackage{etoolbox}
-\usepackage{comment}
+% \usepackage{comment}
 \usepackage{minted}
 
 \setminted[Julia]{
