### 2019-09-23 #44 @MikeInnes
Sorry, not familiar with Whedon so don't know a more efficient way than just removing packages it doesn't like one by one :)
