### 2019-09-23 #44 @MikeInnes
Well, we're using minted for syntax highlighting; should we be doing something different there?
