### 2019-09-23 #44 @vchuravy
Normally you can add the `sty` files you need yourself. But minted requires shelling out to python and I am afraid the environment doesn't support that.
We recommend listings for highlighting.
