### 2019-09-25 093528d
Update paper.tex
