+++ paper/paper.tex
@@ -12,13 +12,16 @@
 \usepackage{amsmath}
 \usepackage{etoolbox}
 % \usepackage{comment}
-\usepackage{minted}
+% \usepackage{minted}
+\usepackage{listings}
 
-\setminted[Julia]{
-    fontfamily=fdm,
-    fontsize=\footnotesize,
-    escapeinside=||,mathescape=true
-    }
+% \setminted[Julia]{
+%     fontfamily=fdm,
+%     fontsize=\footnotesize,
+%     escapeinside=,mathescape=true
+%     }
+
+\lstset{language=Julia,mathescape=true}
 
 \newtoggle{anonymous}
 % \toggletrue{anonymous}
@@ -84,7 +87,7 @@ One feature of our example is that the number of terms will not be fixed, but wi
 
 To run, install Julia v1.1 or higher, and install the Zygote.jl and ForwardDiff.jl packages with:
 
-\begin{minted}{Julia}
+\begin{lstlisting}
 using Pkg
 Pkg.add("Zygote")
 Pkg.add("ForwardDiff")
@@ -104,11 +107,11 @@ function s(x)
     end
     return t 
 end
-\end{minted}
+\end{lstlisting}
 
 While the Taylor series for sine could have been written more compactly in Julia, for purposes of illustrating more complex programs, we purposefully used a loop, a conditional, a print statement, and function calls to \verb+isodd+ and \verb+factorial+, which are native Julia implementations. AD just works, and that is the powerful part of the Julia approach. Let’s compute the gradient at x = 1.0 and check whether it matches \verb+cos(1.0)+:
 
-\begin{minted}{Julia}
+\begin{lstlisting}{Julia}
 julia> # Forward Mode AD
 
 julia> ForwardDiff.derivative(s, 1.0)
@@ -131,7 +134,7 @@ i=11
 
 julia> cos(1.0)
 0.5403023058681398 
-\end{minted}
+\end{lstlisting}
 
     
 \section{Implementation}
@@ -167,17 +170,17 @@ i.e. $\mathcal{J}(f)(x)$ returns the value of $f$ at x, as well as a function wh
 
 \begin{figure}
     \begin{minipage}{0.5\textwidth}
-    \begin{minted}[
+    \begin{lstlisting}[
     framesep=2mm,
     ]{Julia}
-    function |$\mathcal{J}$|(|$f \circ g$|)(x)
-        |$a$|, |$\underline{a}$| = |$\mathcal{J}(f)(x)$|
-        |$b$|, |$\underline{b}$| = |$\mathcal{J}(g)(a)$|
-        |$b$|, |$z$| |$\to$| begin
-            |$\underline{a}$($\underline{b}$(z))|
+    function $\mathcal{J}$($f \circ g$)(x)
+        $a$, $\underline{a}$ = $\mathcal{J}(f)(x)$
+        $b$, $\underline{b}$ = $\mathcal{J}(g)(a)$
+        $b$, $z$ $\to$ begin
+            $\underline{a}$($\underline{b}$(z))
         end
     end
-    \end{minted}
+    \end{lstlisting}
     % \begin{verbatim}
     % function J(f . g)(x)
     %     a, da = J(f)(x)
@@ -189,12 +192,12 @@ i.e. $\mathcal{J}(f)(x)$ returns the value of $f$ at x, as well as a function wh
     \end{minipage}
     \hspace{0.1in}
     \begin{minipage}{0.5\textwidth}
-    \begin{minted}[
+    \begin{lstlisting}[
     framesep=2mm,
     fontsize=\footnotesize,
     fontfamily=fdm,
     xleftmargin=6pt,
[…]
