+++ paper/paper.tex
@@ -192,13 +192,15 @@ i.e. $\mathcal{J}(f)(x)$ returns the value of $f$ at x, as well as a function wh
     \end{minipage}
     \hspace{0.1in}
     \begin{minipage}{0.5\textwidth}
-    \begin{lstlisting}[
-    framesep=2mm,
-    fontsize=\footnotesize,
-    fontfamily=fdm,
-    xleftmargin=6pt,
-    escapeinside=,mathescape=true
-    ]{Julia}
+    % \begin{lstlisting}
+    % [
+    % framesep=2mm,
+    % fontsize=\footnotesize,
+    % fontfamily=fdm,
+    % xleftmargin=6pt,
+    % escapeinside=,mathescape=true
+    % ]{Julia}
+    \begin{lstlisting}
     julia> f(x) = x^2 + 3x + 1
     julia> gradient(f, 1/3)
     (3.6666666666666665,)
@@ -288,8 +290,8 @@ Layers & Total Overhead & Number of Ops & Overhead/op \\ \hline
 2      & 280.5 us & 491 & 571.3 ns \\ 
 3      & 406.1 us & 727 & 558.6 ns \\ \hline
 \end{tabular}
-\label{table:zygote_overhead}
 \caption{\Zygote\ per-op overhead estimated across varying numbers of stacked LSTM layers.}
+\label{table:zygote_overhead}
 \end{table}
 
 This vanishing overhead lowers the bar for AD systems to be integrated more thoroughly into programming languages at an extremely fine scale without worrying about performance issues. The lower the AD overhead, the smaller the minimum feasible kernel size for an AD system that is restricted by backwards pass efficiency.
@@ -431,59 +433,59 @@ We use a standard differentiable variational quantum circuit composed of layers 
 
 The analytical formula for the adjoint of the strong solution of a SDE is difficult to efficiently calculate due to the lack of classical differentiability of the solution\footnote{The the strong solution of stochastic differential equations are non-differentiable except on a measure zero set with probability 1. This means that standard Newtonian calculus cannot hold, and thus the derivation must take place in a the setting of a stochastic functional calculus known as the Malliavin Calculus \cite{doi:10.1137/S0363012902419059,huang2000malliavin,zhang_2004}.}. However, \Zygote\ still manages to calculate a useful derivative for optimization with respect to single solutions by treating the Brownian process as fixed and applying forward-mode automatic differentiation, showcasing \Zygote's ability to efficiently optimize its AD through mixed-mode approaches \cite{2018arXiv181201892R}. Common numerical techniques require computing the gradient with respect to a difference over thousands of trajectories to receive an average cost, while our numerical experiments suggest that it is sufficient with \Zygote\ to perform gradient decent on a neural SDE using single trajectories, reducing the overall computational cost by this thousands. This methodological advance combined with GPU-accelerated high order adaptive SDE integrators in DifferentialEquations.jl makes a whole new field of study accessible. 
 
-\begin{comment}
+% \begin{comment}
 
-\subsection{Combining Probabilistic Programming with Photo-Realistic Computer Vision}
-\label{comp-vis}
+% \subsection{Combining Probabilistic Programming with Photo-Realistic Computer Vision}
+% \label{comp-vis}
 
-Here we show how to construct custom \texttt{UniformRGB} and \texttt{NormalRGB} distributions and their log PDFs. We combine this with the Turing probabilistic programming language, defining a generative model of an image where $\mu \sim \textnormal{UniformRGB}()$ and each $\textnormal{pixel} \sim \textnormal{NormalRGB}(\mu, \sigma)$.
[…]
