+++ paper/ref.bib
@@ -1043,7 +1043,7 @@ eprint = {
   organization={ACM}
 }
 
-@artcle{2019sbc,
+@article{2019sbc,
   title={{J}ulia {E} {F}lux: {M}odernizando O {A}prendizado De {M}áquina},
   author={Gandhi, Dhairya and Innes, Michael and Saba, Elliot and Fischer, Keno and Shah, Viral},
   url={https://www.sbc.org.br/component/flippingbook/book/43/1?page=41},
