### 2019-09-25 #44 @MikeInnes
Alright, that was a journey, but we got there.
