### 2019-10-16 #55 @whedon
[REVIEW]: ∂P: A Differentiable Programming System to Bridge Machine Learning and Scientific Computing
**Submitting author:** @MikeInnes (<a href="http://orcid.org/0000-0003-0788-0242">Michael Innes</a>)
**Repository:** <a href="https://github.com/MikeInnes/zygote-paper" target ="_blank">https://github.com/MikeInnes/zygote-paper</a>
