### 2019-10-16 MikeInnes/zygote-paper#13 @tlienart
[review] License missing
The Whedon checklist https://github.com/JuliaCon/proceedings-review/issues/55 kindly requires some [OSI approved license](https://opensource.org/licenses/alphabetical).
