### 2019-10-16 #55 @tlienart
License missing -- https://github.com/MikeInnes/zygote-paper/issues/13
