### 2019-10-16 #55 @jiahao
Disclosing potential COI with one author (CR) - we are writing a manuscript for publication (not submitted yet) - otherwise no other COIs to declare. (The PNAS guidelines make no mention of future co-authorship).
