### 2019-11-04 MikeInnes/zygote-paper#13 @MikeInnes
Closed by https://github.com/MikeInnes/zygote-paper/commit/9f1ecc895b2ab825b5d4c0bdca746d9660117d11
