### 2020-07-27 #55 @vchuravy
According to a conversation on Slack this paper has been published elsewhere.
